+++ src/pages/users/[userId]/index.js
@@ -1,6 +1,6 @@
 import { useSession,signOut } from 'next-auth/react';
 import { useRouter } from 'next/router';
-import CompetitionItem from '../../components/users/CompetitionItem';
+import CompetitionItem from '../../../components/users/CompetitionItem';
 import { Redirect } from "react-router-dom";
 
 
@@ -9,10 +9,8 @@ function userIndex({data}){
   
   const router = useRouter()
   const { data: session } = useSession()
-  /* state = {
-
-  } */
-  if(session && session.username == router.query.users){
+  
+  if(session && session.username == router.query.userId){
     console.log('Amogus')
     return(
       <div>
@@ -44,39 +42,10 @@ function userIndex({data}){
  
 }
 
-/* export async function getServerSideProps(context) {
-  let config = {
-    method: 'POST',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json'
-    },
-    body: {user : context.params.users}}
-    console.log(config.body)
-    const testData = await fetch("http://localhost:3000/api/userIndex",config)
-    //console.log(testData)
-    const r = await testData.json()
-
-  return {
-    props: {data : r}, 
-  }
-} */
-
 userIndex.getInitialProps = async (ctx) => {
-
-     let config = {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json'
-      },
-      body:{user: ctx.query.users}}
-      console.log(config.body)
-
       const testData = await fetch("http://localhost:3000/api/userIndex")
       const r = await testData.json()
       return {data : r}
-    
   }
 
 export default userIndex
